+++ check/verify.sh
@@ -1,3 +1,4 @@
 #!/bin/bash
 ./merge.py
 mypy tko --check-untyped-defs
+rm tko